requirements-py.txt
ansible==9.7.0
PyYAML==6.0.1
